AutoPlanAgent/GPT-5-2-Codex/pyproject.toml
[project]
name = "autoplan-agent"
version = "0.1.0"
description = "Autonomous planning data analysis agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langchain>=1.0.7",
  "langchain-openai>=1.0.3",
  "langchain-community>=0.4.1",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "pandas>=2.3.3",
  "numpy>=2.1.0",
  "scipy>=1.14.0",
  "scikit-learn>=1.5.0",
  "matplotlib>=3.9.0",
  "seaborn>=0.13.2",
  "pymysql>=1.1.1",
  "jinja2>=3.1.4",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "tavily-python>=0.7.13",
  "trafilatura>=2.0.0",
  "markdownify>=1.2.2",
  "reportlab>=4.2.5",
]
